requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Cors==4.0.0
Flask-Session==0.8.0
Authlib==1.3.1
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
openai==1.35.1
langchain==0.2.11
langchain-openai==0.1.16
celery==5.5.3
redis==5.0.7
requests
python-slugify==8.0.4
gunicorn==22.0.0
resend
PyJWT==2.10.1
alembic==1.13.1